Generate an arbitrary number of random samples from a normal distribution

Returns a data frame whose columns are random samples from a normal distribution

randomSample(mu = 0, sigma = 1, sampleSize = 25, numSamples = 1)

Arguments

mu
mean of normal distribution

sigma
standard deviation of normal distribution

sampleSize
size of each sample

numSamples
number of random samples to be generated

Value

A data frame whose columns are random samples from a normal distribution

Examples

head(randomSample(mu=1,sigma=4,sampleSize=100,numSamples=5))

#>           X1          X2         X3         X4        X5
#> 1  3.4235112 -3.43984862  2.3331666 -0.2498656  3.995852
#> 2 -0.2011225  1.59762278  6.9115645 11.0950479 -7.071990
#> 3  5.0419697 -0.54457094  0.9554800 -0.6788411 -5.248512
#> 4 -6.0775329  0.07802311 -1.7139053  6.3487764  3.583830
#> 5  3.6888107  2.56842379 -0.4342616  0.1070977  2.945809
#> 6 -2.3050294  2.87973270  3.4064665  3.3724917  0.609401